backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Werkzeug==3.0.4
gunicorn==23.0.0
